refactor(site): move layout.shared out of lib/
lib/ is the non-visual layer (i18n, meta, source, codecs), but
layout.shared.tsx rendered components - it assembles the LanguageSwitch
into the Fumadocs nav options. It is docs-layout UI configuration, so it
lives in components/ as docs-layout-options.tsx (named after its export,
baseOptions). Single consumer updated; no behavior change.
